requirements.txt
africastalking
Flask
google-generativeai
logging
